Update bundreport.cls
Added the first of (what I presume will be) many compatibility patches to ensure smooth conversion between Pandoc and LaTex:

Example - when writing a list in markdown, Pandoc automatically formats the list with \begin{itemize} BUT also adds a command \tightlist under it. If this command doesn't exist in the .cls, then the compilation will crash. Hence this commit includes a patch that fixes this issue by simply creating the \tightlist command in the .cls. I presume we will have many issues with this in the future...